updated registerClient
