+++ src/controller/agentLedgerController.js
@@ -9,14 +9,11 @@ export const getAgentTransactions = async (req, res) => {
     const { limit = 30, offset = 0 } = req.query;
     const agentId = req.session.userId;
 
-    // Sanitize limit and offset
     const recordsLimit = Math.min(Math.max(parseInt(limit) || 30, 1), 100);
     const recordsOffset = Math.max(parseInt(offset) || 0, 0);
 
-    // Get filtering conditions from utility
     const conditions = filterUtils(req.query);
 
-    // Fetch agent's commission rates
     const [agent] = await db
       .select({
         maxCasinoCommission: agents.maxCasinoCommission,
@@ -26,7 +23,6 @@ export const getAgentTransactions = async (req, res) => {
       .from(agents)
       .where(eq(agents.userId, agentId));
 
-    // Fetch transactions with calculated fields
     const results = await db
       .select({
         agentId: users.id,
@@ -40,7 +36,6 @@ export const getAgentTransactions = async (req, res) => {
         lossAmount: ledger.debit,
         credit: ledger.credit,
         debit: ledger.debit,
-        // Calculate commission based on game type and transaction type
         agentCommission: sql`CASE 
           WHEN ${ledger.entry} LIKE '%casino%' THEN ${ledger.stakeAmount} * ${agent.maxCasinoCommission} / 100
           WHEN ${ledger.entry} LIKE '%lottery%' THEN ${ledger.stakeAmount} * ${agent.maxLotteryCommission} / 100
@@ -60,7 +55,6 @@ export const getAgentTransactions = async (req, res) => {
       .limit(recordsLimit)
       .offset(recordsOffset);
 
-    // Calculate running totals
     let runningBalance = 0;
     let totalCommission = 0;
     let totalProfit = 0;
@@ -83,7 +77,6 @@ export const getAgentTransactions = async (req, res) => {
       };
     });
 
-    // Fetch total record count
     const totalRecords = await db
       .select({ count: sql`COUNT(*)` })
       .from(ledger)
@@ -92,7 +85,6 @@ export const getAgentTransactions = async (req, res) => {
       .innerJoin(agents, eq(players.agentId, agents.id))
       .where(and(eq(agents.userId, agentId), ...conditions));
 
-    // Calculate next offset
     const nextOffset =
       recordsOffset + recordsLimit < totalRecords[0].count
         ? recordsOffset + recordsLimit
@@ -103,7 +95,7 @@ export const getAgentTransactions = async (req, res) => {
       success: true,
       message: "Agent transactions fetched successfully",
       data: {
-        transactions: formattedResults,
+        results: formattedResults,
         summary: {
           totalProfit,
           totalLoss,
@@ -117,11 +109,7 @@ export const getAgentTransactions = async (req, res) => {
       },
     };
 
-    logToFolderInfo(
-      "Transactions/controller",
-      "getAgentTransactions",
-      response
-    );
+    logToFolderInfo("Transactions/controller", "getAgentTransactions", response);
     return res.json(response);
   } catch (error) {
     let errorResponse = {
@@ -131,11 +119,7 @@ export const getAgentTransactions = async (req, res) => {
       error: error.message,
     };
 
-    logToFolderError(
-      "Transactions/controller",
-      "getAgentTransactions",
-      errorResponse
-    );
+    logToFolderError("Transactions/controller", "getAgentTransactions", errorResponse);
     return res.status(500).json(errorResponse);
   }
 };
@@ -158,7 +142,6 @@ export const createTransactionEntry = async (req, res) => {
   try {
     const agentId = req.session.userId;
 
-    // Verify agent owns this player
     const [player] = await db
       .select()
       .from(players)
@@ -191,21 +174,20 @@ export const createTransactionEntry = async (req, res) => {
       result: note,
     };
 
-    // Insert transaction
     const result = await db.insert(ledger).values(newEntry);
 
     let response = {
       uniqueCode: "CGP0083",
       success: true,
       message: "Transaction entry created successfully",
-      data: { results: result },
+      data: {
+        results: [result], 
+        summary: {},
+        pagination: {},
+      },
     };
 
-    logToFolderInfo(
-      "Transactions/controller",
-      "createTransactionEntry",
-      response
-    );
+    logToFolderInfo("Transactions/controller", "createTransactionEntry", response);
     return res.status(201).json(response);
   } catch (error) {
     let errorResponse = {
@@ -215,11 +197,7 @@ export const createTransactionEntry = async (req, res) => {
       error: error.message,
     };
 
-    logToFolderError(
-      "Transactions/controller",
-      "createTransactionEntry",
-      errorResponse
-    );
+    logToFolderError("Transactions/controller", "createTransactionEntry", errorResponse);
     console.error("Error creating transaction entry:", error);
     return res.status(500).json(errorResponse);
   }
