+++ src/routes/clientRoutesAPI.js
@@ -26,7 +26,7 @@ import {
 } from "../controller/gameController/index.js";
 import { checkBlockingLevel } from "../middleware/checkBlockingLevel.js";
 import { changePassword } from "../controller/passwordController/index.js";
-import { getClientLedger } from "../controller/clientLedgerController.js";
+import { getClientLedger } from "../controller/clientLedger/clientLedgerController.js";
 import { toggleFavoriteGame } from "../controller/favouriteGameController/toggleFavoriteGame.js";
 import { getPlayHistory } from "../controller/histroyController/getPlayHistory.js";
 const router = express.Router();
